Update Usage, Debug, Configuration structures

+++ bin/createramdisk.sh
@@ -6,19 +6,20 @@
 # @company Frobas IT Department, www.frobas.com 2015
 # @author  Vladimir Roncevic <vladimir.roncevic@frobas.com>
 #
-UTIL_NAME=createramdisk
+UTIL_CREATERAMDISK=createramdisk
 UTIL_VERSION=ver.1.0
 UTIL=/root/scripts/sh-util-srv/$UTIL_VERSION
 UTIL_LOG=$UTIL/log
 
 . $UTIL/bin/usage.sh
+. $UTIL/bin/checktool.sh
 . $UTIL/bin/devel.sh
 
-declare -A TOOL_USAGE=(
-    [TOOL_NAME]="__$UTIL_NAME"
+declare -A CREATERAMDISK_USAGE=(
+    [TOOL_NAME]="__$UTIL_CREATERAMDISK"
     [ARG1]="[MOUNTPT] Mount point"
     [EX-PRE]="# Example creating RAM disk"
-    [EX]="__$UTIL_NAME \"/mnt/test/\""	
+    [EX]="__$UTIL_CREATERAMDISK \"/mnt/test/\""
 )
 
 #
@@ -30,50 +31,64 @@ declare -A TOOL_USAGE=(
 # @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
 #
 # __createramdisk "$MOUNTPT"
-# STATUS=$?
+# local STATUS=$?
 #
 # if [ "$STATUS" -eq "$SUCCESS" ]; then
 #   # true
+#	# notify admin | user
 # else
 #   # false
+#	# missing argument | missing tool
+#	# return $NOT_SUCCESS
+#	# or
+#	# exit 128
 # fi
 #
 function __createramdisk() {
-    MOUNTPT=$1
+    local MOUNTPT=$1
     if [ -n "$MOUNTPT" ]; then
-        SIZE=2000            # 2K blocks (change as appropriate)
-        BLOCKSIZE=1024       # 1K (1024 byte) block size
-        DEVICE=/dev/ram0     # First ram device
-		if [ "$TOOL_DEBUG" == "true" ]; then
-			printf "%s\n" "[Creating RAM disk]"
-            printf "%s" "Checking directory [$MOUNTPT] "
+		local FUNC=${FUNCNAME[0]}
+		local MSG=""
+		local MKE2FS="/sbin/mke2fs"
+        local SIZE=2000            # 2K blocks (change as appropriate)
+        local BLOCKSIZE=1024       # 1K (1024 byte) block size
+        local DEVICE=/dev/ram0     # First ram device
+		if [ "$TOOL_DBG" == "true" ]; then
+            MSG="Check dir [$MOUNTPT]"
+			printf "$DQUE" "$UTIL_CREATERAMDISK" "$FUNC" "$MSG"
 		fi
         if [ ! -d "$MOUNTPT" ]; then
-			if [ "$TOOL_DEBUG" == "true" ]; then
+			if [ "$TOOL_DBG" == "true" ]; then
 				printf "%s\n" "[not ok]"
-            	printf "%s\n" "Creating directory [$MOUNTPT]"
+            	MSG="Creating directory [$MOUNTPT]"
+				printf "$DSTA" "$UTIL_CREATERAMDISK" "$FUNC" "$MSG"
 			fi
             mkdir "$MOUNTPT"
         fi
-		if [ "$TOOL_DEBUG" == "true" ]; then
+		if [ "$TOOL_DBG" == "true" ]; then
             printf "%s\n" "[ok]"
 		fi
-        dd if=/dev/zero of=$DEVICE count=$SIZE bs=$BLOCKSIZE  
-        mke2fs "$DEVICE"
-		if [ "$TOOL_DEBUG" == "true" ]; then
-			printf "%s\n" "Mounting device [$DEVICE] to point [$MOUNTPT]"
+		__checktool "$MKE2FS"
+		local STATUS=$?
+		if [ "$STATUS" -eq "$SUCCESS" ]; then
+			dd if=/dev/zero of=$DEVICE count=$SIZE bs=$BLOCKSIZE  
+			eval "$MKE2FS $DEVICE"
+			if [ "$TOOL_DBG" == "true" ]; then
+				MSG="Mounting device [$DEVICE] to point [$MOUNTPT]"
+				printf "$DSTA" "$UTIL_CREATERAMDISK" "$FUNC" "$MSG"
+			fi
+			mount "$DEVICE" "$MOUNTPT"
+			if [ "$TOOL_DBG" == "true" ]; then
+				printf "$DSTA" "$UTIL_CREATERAMDISK" "$FUNC" "Set permission"
+			fi
+			chmod 777 "$MOUNTPT"
+			if [ "$TOOL_DBG" == "true" ]; then
+				printf "$DEND" "$UTIL_CREATERAMDISK" "$FUNC" "Done"
+			fi
+			return $SUCCESS
 		fi
-        mount "$DEVICE" "$MOUNTPT"
-		if [ "$TOOL_DEBUG" == "true" ]; then
-        	printf "%s\n" "Set permissions"
-		fi
-        chmod 777 "$MOUNTPT"
-		if [ "$TOOL_DEBUG" == "true" ]; then
-        	printf "%s\n\n" "[Done]"
-		fi
-        return $SUCCESS
+		return $NOT_SUCCESS
     fi
-    __usage $TOOL_USAGE
+    __usage $CREATERAMDISK_USAGE
     return $NOT_SUCCESS
 }
-
